+++ govtool/backend/src/VVA/API.hs
@@ -244,13 +244,13 @@ listProposals selectedTypes sortMode mPage mPageSize mDrepRaw mSearchQuery = do
 
 
 
-  let filterF ProposalResponse{..} = case mSearchQuery of
+  let filterF ProposalResponse{..} = case Text.toLower <$> mSearchQuery of
         Nothing -> True
         Just searchQuery -> fromMaybe False $ do
-          title <- proposalResponseTitle
-          about <- proposalResponseAbout
-          motivation <- proposalResponseMotivation
-          rationale <- proposalResponseRationale
+          title <- Text.toLower <$> proposalResponseTitle
+          about <- Text.toLower <$> proposalResponseAbout
+          motivation <- Text.toLower <$> proposalResponseMotivation
+          rationale <- Text.toLower <$> proposalResponseRationale
 
           let result = searchQuery `isInfixOf` title
                       || searchQuery `isInfixOf` about